Make ConvParams for sub-convolutions as we iterate the partition.
Summary:
This change also affects how input and weight tensors are mapped. Call
getElementIndices() to, given a slice read by a tile, get the regions
in the original (untransformed) tensor and use this to calculate a tile
mapping for the original tensor.

Fixes T2234
Depends on D3558

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T2234

Differential Revision: https://phabricator.sourcevertex.net/D3561