Express the partial cycles as function of conv size variables.
Summary:
This avoids re-computation of the size using the various getMaxTileSize
functions (which can be removed). I tried to make this change
previously around the time 3ed9dbec landed but abandoned it due to an
increase in the popsolver runtime. I've addressed this by introducing a
dummy call that takes the split variables as arguments so the order that
popsolver sets the variables is not affected. I've left a TODO about
finding a less hacky way of achieving the same result.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3638