Allow transforms at any level of the hierarchy.
Summary:
Previously transforms were always applied at the top level before
splitting the convolution. After this change transforms may be performed at
any level. Transforms are interleaved with the splitting into
sub-convolutions.

The planner doesn't yet take full advantage of this new found freedom,
in particular the planner is hardcoded to perform transforms at the IPU
level, with the exception of swapping operands / adding dimensions which
is always performed at the top level.

As part of this change I've removed the inChansPadding /
partialChansPadding fields from ConvTransform. After the convolution is
split there may be a different number of input channels on each tile and
therefore the amount of padding required may be different on different
IPUs. Instead of looking at the padding field we implicitly pad to a
multiple of the grain size.

Fixes T2239

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T2239

Differential Revision: https://phabricator.sourcevertex.net/D3683