Add support for different partial types at each level of the hierarchy.
Summary:
At each level there is a partial type that controls the type used for
intermediate calculations and a result type that controls the type of
the result produced at that level. For now the planner always sets the
the result type is set to the partial type, except for at the top level
where it is set to the result type of the convolution.

Depends on D3683

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D3689