Force the void fns used as type-ids to be truely singletons
Summary:
when tensorflow was using the new Expr code, it was failing to
pass the type identification code in inferType.  This seems to
be because the type-id function in ExprType is instatiated
inline, and therefore exists in both the poplar DSO and the TF
DSI files.  I'm not sure why weak runtime linking doesn't
resolve this - although I can see that the runtime linker has
a difficult decision to make when presented with two symbols
which are both in DSO files that appear to need resolving weakly.

Test Plan: all existing tests running.  TF now ok.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3743