Fix support for arbitrary output channel groupings.
Summary:
The weight update of a fully connected layer uses an output channel
grouping of 16 - this requires two passes per output group. The cycle
estimation for this case was broken in 26d89f36 when we moved to the new
vertex memory layout.

Prior to 26d89f36 it was easy to support multiple passes without changing
the vertex code by setting the output pointers, weight pointers and the output
stride appropriately. We no longer have a weight pointer per pass and so
some changes will be required to the codelet. For now I've only updated the
cycle count, I haven't updated the vertex state.

Since we will need changes in the codelet to support multiple passes
per group this I've also assumed can make the same change to the 1x1 vertex.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D3795