Update the planner to try output channel groupings of 16.
Summary:
Using an output channel grouping of 16 avoids the need to regroup the
output between this layer and the next and it increases the cross-tile
block size used when weights are rearranged between the forward and
backward passes. Update the planner to try an output channel grouping of
16 and update addTransformCycleEstimate() to add a small penalty to plans
that don't use an output grouping of 16.

For ResNet-18 training the effect is as follows:

| Batch size | Mem before | Mem after | Cycles before | Cycles after |
| 1          | 216.9      | 214.6     |  697551       | 701361       |
| 4          | 338.0      | 338.2     |  1732274      | 1732347      |

For ResNet-50 training the effect is as follows:

| Batch size | Mem before | Mem after | Cycles before | Cycles after |
| 1          | 336.4      | 327.5     | 1703386       |  1712017     |

Fixes T2570.

Depends on D3795

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T2570

Differential Revision: https://phabricator.sourcevertex.net/D3796