Optimize 1x1 dilated convolutional layers
Summary:
Given a dilated 1x1 convolutional layer we can commute the dilation and
the convolution operation, applying the dilation after the partial
sums have been reduced. This reduces the size of the partials variable,
reducing memory usage and the amount of exchange / reduction.

The dilatePostConv member of the plan controls whether this optimization
is applied. The planner always applies this optimization if it is legal
to do so.

On ResNet-50 training this reduces the number of cycles from 1712017
to 1691747 and it reduces the max live tensor data from 113.1 MB to
104.3 MB. The overall memory usage increases 327.6 MB 330.2 MB but the gaps
in the variable data also increases from 29.2 MB to 39.8 MB so in future
when we start filling gaps with exchange code / improve the storage allocator
I'd expect the memory usage to improve.

Fixes T2611

Depends on D3808

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T2611

Differential Revision: https://phabricator.sourcevertex.net/D3809