Base whether we set BOOST_TEST_DYN_LINK on the library suffix.
If Boost_USE_STATIC_LIBS is OFF FindPackage(Boost) searches for both
shared libraries and static libraries (in that order) and so
Boost_USE_STATIC_LIBS can't be used to determine the library type.

The best (or least worst) way of detecting the library type appears to be
by checking the extension - that's what I've implemented here.