Minor batch normalisation documentation update.
Summary:
There is a function batchNormalise which returns a std::pair<Tensor, Tensor>.
The FIRST tensor is the output of the layer, with gamma and beta applied to it.
The comment to the function stated that the output of the layer was the SECOND.
This commit reorders the comment.

Maybe in the future we should not return std::pair, but an object with sensibly
named class members.

Test Plan: No tests as no change to source.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D3992