Modify channel processing operations in popconv to use cntiguous field per channel group
Summary:
The existing method creates an edge per batch and channel group
even though the layout keeps the batch and field dimension
together. In all existing benchmarks it is seen that a single
output group is mapped to a tile. A specialisation for this case
is added with a supervisor vertex dividing up the contiguous
field. The original method is retained if more than one channel
group is assigned to a tile. This can be removed in future and
a single supervisor vertex with multiple edges connected to
a contiguous field.

Test Plan: All single layer tests pass

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T2892

Differential Revision: https://phabricator.sourcevertex.net/D4103