Constrain tile splits to the smallest divisors that give that result.
Summary:
Previously if we wanted to divide an axis we would try every possible
divisor from 1 to the size of the axis. When two different divisors give
the same result then there is no point exploring both options - the
number of cycles depends on the tile with the most amount of work to do
which will be the same in both cases (if everything else is equal). Add
constraints to ensure the divisor we pick is the smallest divisor that
gives that result. This cuts down the size of search space the planner
must explore, speeding up the planner.

For ResNet-18 training batch size 2 this reduces the time spent in the
planner from 17.5 seconds to 6.5 seconds. A slightly different plan
is chosen which has the following effect:

Cycles: 1078737 -> 1079733
Total memory: 245MB -> 240MB
Max per-tile memory: 235KB -> 238KB

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D4149