New reduction implementation
Summary:
New reduction code. This isn't intended to be the actual merged code - there are two main changes that need to be done before it is merged, but I think it can be reviewed like this:

* There's a lot of visualisation code (including a copy of Yoga) that shouldn't be in here. I'll remove it before merging and instead in future support serialisation of the ReductionDebug structure.
* I haven't removed the convChannelReduce() and pointed it to this code yet. And actually one use of (batchNormEstimates) it does two reductions in parallel which the API does not support, but I propose for now that it just does them one after the other.
* Probably want to replace the current public API (include/popops/Reduce.hpp) with the one in lib/popops/reduction/Reduction.hpp

I've broken out a lot of the support and util functions into separate commits, and then the final commit is all the new reduction code (sorry). Here are some more limitations:

* Specifying types (i.e. when to use half/float) is not really dealt with properly. I guess this should be done similarly to how it will be done in convolutions.
* There are some potential optimisations and magic numbers in the code (search for "Optimisation:")
* I haven't added the ABS_ADD and SQUARE_ADD operations (easy to do though)
* The codelets use an ugly macro, and also partials always uses TWO_VECTOR even when it doesn't need to (if partialsAreOutputSize is true)
* reduce_op is added as a way to do random reduction tests, but I haven't added `add_test()`'s that use it yet
* It doesn't consider multi-IPUs yet. It probably needs to be written in a more generic hierachical way for that, but one step at a time.

See T1588

Test Plan: Tested using random reductions (see reduce_op), and also reductions recorded from resnet50. On average it is a bit slower than convChannelReduce() (though it uses a different cycle estimator so hard to be sure), but it uses much less memory and the compile time is almost always less).

Reviewers: richard, timh

Reviewed By: richard

Subscribers: samk, davel

Differential Revision: https://phabricator.sourcevertex.net/D3838