Make 1D ScaledAdd a supervisor vertex
Summary:
Scaled add is not a major cycles consumer but uses a lot
of vertex state. As most of the edges to ScaledAdd for
RESNET50 training are single regions, it can be made into
a supervisor task. This saves 1/6th of vertex state (~5MB
for RESNET50 training)

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T2610

Differential Revision: https://phabricator.sourcevertex.net/D4182