Fix planning for fully connected layer with an input size of 1.
Summary:
If the input size is one the planner uses the outer product method in the
forward pass. Previously it used the same method in the backward and
weight update methods, despite the fact that the requirements for using
the output product method may not be met in these passes.

Fixes T2934

Reviewers: davidn

Reviewed By: davidn

Maniphest Tasks: T2934

Differential Revision: https://phabricator.sourcevertex.net/D4181