Optimize ConvUtils functions to avoid looping.
Summary:
Add functions to quickly compute input / output / kernel ranges based
on the bounds of the range instead of looping over every element.
Keep track of whether the range is exact (i.e. matches the true
range) or whether it is an conservative approximation. If the
range is an approximation we fall back to the old way of computing the
true range. In the majority of cases there is no need to loop over
every element which significantly speeds up these functions. Even when
the range is not exact it can often be used to reduce the number of
element we need to loop over.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D4159