Make single edge non-linearity codelets supervisor vertices
Summary:
If the output edge is connected to a single variable, all the
regions connected to the edges are gathered into a single region.
A supervisor vertex is used to divide the work.

This results in decrease in vertex state but an increase in
total compute cost. This cost can be reduced by setting up
scaled pointers in the supervisor rather than the worker. The
cycle cost estimates make an assumption that the scaled pointers
are expanded to physical addresses in the worker.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D4328