Create the fewer convolution output tensors where possible
Summary:
As a convolution graph is created it recurses through a heirarchy of
sub-convolutions. It is more efficient to create one output variable for
sevreal sub-convolutions than create many and stiching them together.
This change tries to create such a variable as early as possible.

No functionality change

This is a re-application of commit 52f36430d8ac453070ad5b8e82192d53f7f7e with
fixes.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D4469