Add codelets for in-place unary, binary and ternary operators
Summary:
The existing implementation used the normal operations which
created the output tensor which was copied into the first input
tensor. Codelets are created and specialisations are defined
only for operations wherein the output tensor type is the same
as the input type.

The cycle estimates haven't changed because load/st
instructions to interleaved memory where the ld/st is to the
buffer can be done by using two instructions in the inner loop.
The pair of instructions utilises double the number of registers
required.

AddTo and subtractFrom without any scaling factor are replaced
by the inplace form.  scaledAddTo and scaledSubtractFrom replace
addTo and subtractFrom with a scaling factor.

Test Plan: New test added for ternary inplace operation

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T3017

Differential Revision: https://phabricator.sourcevertex.net/D4299