New Loss Layer implementation (and Loss tests)
Summary:
Fixes T2929.
Add a new Loss layer implementation that distributes calculation of the loss and deltas across tiles. At this stage this is distributed based purely on the original tile mapping of the input activations which works well for enigma (where these tensors are linearly mapped), but may not be optimal in general.
This commit also splits off the 'Accuracy' calculation (`numCorrect`) into its own layer which currently works similarly to before (max class for each batch is sought on a single tile). A suggestion for a better implementation I experimented with is commented.
Tests for Loss layer and layers it depends on have also been added.

Test Plan: `./test.sh poplibs`, `./test.sh enigma`

Reviewers: davel, godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T2929

Differential Revision: https://phabricator.sourcevertex.net/D4400