Fix finding of codelet files
Summary:
Commit 76c91977a86700a679610 refactored the code to find codelets into a
common place. However, since the search uses the location of the
currently loaded library this changed the behaviour of the search.

This fix passes a location in the current library from the callee into
the common code so that the behaviour is that same as before the
refactoring.

Test Plan: Ran tests with no install folder present.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Subscribers: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D4872