Use single thread for executing bias and batch norm reductions
Summary:
This is a temporary solution to avoid sub-word writes to potentially
overlapping memory. A proper solution is to create tensors and/or the
amount of work assigned to a worker to be at least a word. These
solutions are not explored because this code will be removed and
replaced by the new reductions.

Test Plan: Tested that some failing tests pass on the simulator

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D4895