Use correct activation slices to addToChannel 1D vertex
Summary:
The 1D addToChannel vertex assumed that channel groups of an
activation are mapped to one tile. This is not the case when
the number of activations is very small compared to the number
of tiles.

Use the correct slices by doing a gather if channels per
groups are spread over more than one tile

Test Plan: All poplibs tests pass and failing examples of single conv layer pass

Reviewers: timh, Eddy

Reviewed By: Eddy

Maniphest Tasks: T3511

Differential Revision: https://phabricator.sourcevertex.net/D4936