Assembly implementations of all NonLinearitySupervisor vertex template instantiations.
Summary:
Fixes T3324.
Add assembly implementations for `popnn::NonLinearitySupervisor`.
This changes the vertex state such that the `n` field for element count stores a packed size with elements pre-divided by (number of workers * elements per 64-bits), and the remainder of this division.
For this reason we also now fallback to the 2D vertex if the target has anything but 6 workers currently as we have specialised the vertex for this number.

The implementation is an adaptation of that found in the single worker microbenchmark for 6 workers, without the requirement of being able to potentially read and perform sigm/max/tanh calculations without causing exceptions on the 64-bits past the end of the region.

Added tests to cover different starting/ending alignments of the regions to perform the non-linearity on.

The assembly particularly for the `half` specialisations could be simplified by using a simpler worker vertex if the number of elements is below a certain threshold.
This would make sense anyway as with so few elements the use of all workers is not really necessary.

Test Plan: `./test.sh poplibs` with Sim test variants enabled. Some failures in conv tests which appear to have been present prior to this change and should not be related.

Reviewers: samuelh, Eddy, godfrey.da.costa, samk

Reviewed By: godfrey.da.costa, samk

Subscribers: samk

Maniphest Tasks: T3324

Differential Revision: https://phabricator.sourcevertex.net/D4820