Disable Sim:cpp variants for the NonLinearitySupervisor codelet tests for now.
Summary: Ubuntu/Debug test is failing for `Sim_cpp_NonLinearitySupervisor_half_tanh`. A brief investigation seems to indicate that somehow on Linux the generated code for the codelet is different to Mac, using `$m7` with a garbage value as the stack pointer in a function call and is causing a rather wild exception in the simulator. Needs more investigation but disable this for now as I don't think there's anything particularly wrong with the commit or tests as they are.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D4956