Fix and enable BigNLVertices test
Summary:
Splits the test in to sizes that should and shouldn't fit and
test for the appropriate exception for sizes that do not fit
on tow tiles.

Also simplify the use of VARIANT so that all IpuModel;Sim;Sim:cpp
are enabled by default and exceptions that only run on IpuModel
are explicit.

Reviewers: Eddy, richard, samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D5089