Changes to pass Batch Norm tests on Sim device
Summary:
Change the arithmetic used in the batch norm estimator to match
machine ISA. This also improves the accuracy of the estimates.
Allow the possibility of using a different precision for the squared
reduction used in the power estimator as the dynamic range is higher.
The selection of the type will depend on results obtained from experiments.
The current selection is the same as the activations type to reduce
memory usage.

The tests are modified to fit device memory and the learning rate
adjusted to fit the input data range and type.

Maniphest Task: T3484

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D5092