Ensure that zeroing of partials in convReduce does not produce sub-word writes
Summary:
The 2D tensor created for partials in convChannelReduce could have
it's element zeroed at odd offsets. This results in sub-word writes.
One option is to always make the number of elements per worker to
be a multiple of the atomic store granularity.

The other option would be to create a variable for each worker, which
guarantees that each worker gets an aligned variable. This code is
going to be removed soon once the new reductions are used.

Test Plan: Failing convolution test conv_random_40 passes on simulator

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D5148