Add AMP based convolution codelets for half input type and float output type to poplibs
Summary:
The codelets are enabled for Sim and Hw targets. The supervisor code
needs to be optimised further to reduce bubbles introduced due to
register dependencies.

The code for zeroing partials reduces code bloat but increases compute
cost compare to a specialised version which assumes the partials to be
zero. There are a maximum of 3 additional 64-bit non-strided writes
which needs to be accounted for in the storage allocator.

Test Plan:
All gemm, rnn/lstm, random conv, directed conv, fully connected tests
pass on Sim and Hw targets

Reviewers: richard, Eddy

Reviewed By: Eddy

Subscribers: jamesn

Maniphest Tasks: T3171, T3318, T3316

Differential Revision: https://phabricator.sourcevertex.net/D4791