Change the heuristic for compareTileMapDistributions
Summary:
Current heuristic does not work well in the case where a tensor which occupies more tiles is padded with a constant, meaning that there is a lot of memory regions on Tile 0.
When this tensor is then cloned and an operation is applied to it, the other tensor has to do a lot of copies to get the data into the right place.
(This happens in Resnet where we pad.)

The heuristic is changed to take the tensor with minimum of maximum per tile memory size.

Test Plan: `test.sh poplibs` and `enigma` pass, TF Resnet speeds up.

Reviewers: godfrey.da.costa, davidn

Reviewed By: davidn

Differential Revision: https://phabricator.sourcevertex.net/D5459