Create a file containing utility functions used within popconv
Summary:
There is already a file containing utility functions, namely ConvUtil.cpp
but I didn't want to add more functions there as it would expose them
to the world (ConvUtil.hpp is a public header). I've moved the
functions to convert between different internal formats / shapes to
ConvUtilInternal.cpp along with some miscellaneous functions from
ConvUtil.cpp that weren't used outside of the popconv library.

The reason for this change is I want to call these utility functions
in unit tests - previously this was not possible as the functions
were marked as static.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D5512