Fix scale & update with two-stage reductions on a tile
Summary: If popops had decided to do a two-stage reduction within a tile, the first stage shouldn't do a scale or update.

Test Plan:
After switching from `convChannelReduce` to `popops::reduce`, the MNIST learning is *almost* identical to before.

Before this change:

--- Accuracy after 0 batches = 7.8%
--- Accuracy after 2500 batches = 42.26%
--- Accuracy after 5000 batches = 65.44%
--- Accuracy after 7500 batches = 77.99%
--- Accuracy after 10000 batches = 80.18%

After this change, and using `popops::reduce` instead of `convChannelReduce`

--- Accuracy after 0 batches = 7.8%
--- Accuracy after 2500 batches = 42.26%
--- Accuracy after 5000 batches = 65.44%
--- Accuracy after 7500 batches = 77.97%
--- Accuracy after 10000 batches = 80.6%

The small difference is probably because the reduction does the addition in a different order, e.g. (1 + 2) + (3 + 4) instead of (1 + 2 + 3 + 4).

I also ran MNIST for 2 iterations, and the edge trace results give identical numbers.

Reviewers: Eddy, richard, Richardc

Reviewed By: Richardc

Subscribers: Richardc

Differential Revision: https://phabricator.sourcevertex.net/D5563