Implement AddToChannel2D assembly
Summary: The main change is that the core code of AddToChannel is refactored into a single function, that is called by AddToChannel, and AddToChannel2D. This saves a fair bit of code (~160 instructions in total for the 4 `half` vertices).

Test Plan: AddToChannel2D test is added. It passes. I also test all the `Ipu_default_conv_random` tests which pass.

Reviewers: tomm, samk, samuelh, godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D5477