Added Outer Product assembly vertices and associated test
Summary:
Assembly versions of the Outer Product half and float vertices.
Each has a fast version which takes advantage of 64 bit loads/stores and
vectorised instructions and is expected to cover the cases observed to
be used in resnet50. Also a slower version intended to catch all cases
where the loop size and array size constraints don't allow the fast
version to be used.

Test Plan: Poplar test added.

Reviewers: timh, samk, samuelh, tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D5728