Reimplement Accuracy layer as a special kind of reduction.
Summary:
Fixes T3258.
Accuracy reimplemented as a special type of reduction for the max.
In this case the compute cost per-element is quite high as we have to track the max value and index through.
Also tried to balance the memory cost of needing the index for each element we calculate the max for by keeping the first stage of reduction quite wide with a supervisor vertex and a gathered contiguous region.

This has only really been tuned for the kinds of batch sizes + no. of activations we see in models in benchmarking currently for which there are at most 2 stages of reduction.

Test Plan: `./test.sh poplibs && ./test.sh enigma`

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T3258

Differential Revision: https://phabricator.sourcevertex.net/D5875