Add aseembly implementations of ChannelMul and ChannelMul2D
Summary:
These are very similar to AddToChannel, the main difference being that they are not in-place so you can use memory constraints to put the input and output in different memory elements, and not have to worry about conflicts.
This simplifies the implementation so we don't need the complicated multiple-of-8 pipeline.

There is quite a lot of duplicated code between this and AddToChannel. It could potentially be reduced using assembly macros but I'm not sure it would end up much better.

Also the `float` implementation is not pipelined yet. I have added a task for that:

Test Plan: Test added.

Reviewers: godfrey.da.costa, samuelh, tomm, timp, samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D5656