Split convolution field in 1x1 and nx1 if input strides exceed machine stride
Summary:
For nx1 convolutions, if conv unit weight height is not 1,
multiple kernel positions are used in a single loop for half.
This requires strides which are large if the innermost field
dimensions are large and may not fit machine stride widths.
If this happens, split the field and use as many supervisor
vertices as the split factor. The possibility of output strides
not fitting is very low as it would require very large strides.

For 1x1, the possibility that input and output stride do not
fit is unlikely to happen as it would require a very large
stride (eg it would require a stride of 512/4).

The planner cost is not changed as the overhead cost is
small although utility functions are added in case it needs
to be added.

Maniphest Task: T4180

Test Plan:
single layer tests
conv_flip_input_and_kernel_2
conv_input_dilation_and_stride_share_factor4

test the case when strides exceed machine stride width.
Also tested that RESNET18/batchsize=8/Inf uses more
than 1 vertex for the first layer.

Reviewers: richard, Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D6434