Improve popsys test
Summary:
Test cycleCount method by timing a function that takes a known number of
cycles. Instead of checking for an exact cycle count check the cycle
count is in a sensible range so we don't have to update the test for small
changes in the supervisor code generated by poplar.

Reviewers: samuelh

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D6488