Ensure output of dynamic slice is balanced
Summary:
Currently if a dynamic slice call can product very imbalance
compute/output. The pathalogical case is if the possible slices are not
split across tiles at all. In this case all data is copied to one tile
and sliced to the result!

This change uses the poputills::rebalance to check if the output tensor
is imbalanced and rebalance if necessary.

Depends on D6577

Test Plan: Poplibs test pass + lstm log output is the same (no rebalancing occurs there).

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: davidn, markp, richard

Differential Revision: https://phabricator.sourcevertex.net/D6578