Use constant tensor for number of partials in popops::Reduce rather than variable sized field
Summary:
In the existing implementation it is a variable sized
field element. We don't have a way to specify that
field init state is constant and hence it is better to
make it a constant tensor which can be reused across
multiple vertices if its constituent elements are the
same.

This saves some memory.

Reviewers: samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D6755