ReduceAdd (poplin codelet rewrite), including D6598
Summary:
This includes code copied from the diff:
https://phabricator.sourcevertex.net/D6598
As can be read from that link, the intent is to reduce vertex state by
rearaanging and making supervisor vertices.  The additional work in this diff
is to add the assembly implementation required.

Test Plan:
Run reduceAdd tests, with additional longer tests as code is not excersised
without them

Reviewers: samk, samuelh, timh, tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D6732