Update AMP convolution code to avoid a rearrangement after exchange
Summary:
A rearrangement is required if the input or weights contains aliases.
Avoid creating aliases by creating a variable with size equal to the
amount of padding required and using this in place of zeros.

This has the following effect on ResNet-18:

| Batch | Memory Before  | Memory after   | Time before | Time after |
| 1     | 175.0 MB       | 174.5 MB       | 0.746 ms    | 0.695 ms   |
| 2     | 192.0 MB       | 180.5 MB       | 0.999 ms    | 0.978 ms   |
| 4     | 272.0 MB (OOM) | 254.1 MB       | -           | 1.621 ms   |

and the following effect on ResNet-50:

| Batch | Memory Before  | Memory after   | Time before | Time after |
| 1     | 242.8 MB       | 244.8 MB       | 1.880 ms    | 1.829 ms   |
| 2     | 290.0 MB (OOM) | 290.3 MB (OOM) | -           | -          |

Fixes T4775

Reviewers: Eddy, godfrey.da.costa

Reviewed By: godfrey.da.costa

Subscribers: matt

Maniphest Tasks: T4775

Differential Revision: https://phabricator.sourcevertex.net/D6946