Simplify basicLstmCellForwardPass
Summary:
Previously basicLstmCellForwardPassImpl() acted on sequences but it was
only ever called with a sequence length of 1. Rename
basicLstmCellForwardPassImpl() -> basicLstmCellForwardPass() and
update it to only generate code for one step. Since
basicLstmCellForwardPass() only performs one step it no longer needs
to know which state is going to be retained - it can just return all
the intermediate results and let the caller decide what to keep.

Depends on D6990

Reviewers: tomm, godfrey.da.costa

Reviewed By: tomm, godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D6992