Rearrange dynamic sliced inputs to backpropagation in LSTM.
Summary:
Depends on D7082.  Fixes T4948.

Map retained forward state when training to match mapping we want for dynamic slicing.
We also currently add a rearranging copy of the forward state before doing the backwards pass because we need to append an extra initial state to the sequence in order to dynamicSlice.
This could be removed in a couple of ways but the copy is cheap and simple for now.

Also rearrange the incoming gradients from the next layer.
We should be able to map the gradients output from an LSTM in a stack such that this rearranging copy is elided for the next LSTM down. Currently poplar will likely not currently elide this copy.

Reduces vertex state for DynamicSlice vertices and edges, and due to rebalancing tile workloads in back propagation cuts overall cycles in 2, improving compute percentage for the program as a whole from ~30% to ~60%.

Test Plan: `./test.sh poplibs`

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T4948

Differential Revision: https://phabricator.sourcevertex.net/D7086