Add an inputRHSIsPreArranged option and preArrangeMatMulInputRHS function.
Summary:
preArrangeMatMulInputRHS is a performance optimization available to pre-arrange tensors once
and use the same tensor for multiple fully connected layer matmul operations.
For the training backwards pass this is a transpose of the weight memory for efficiency.

preArrangeMatMulInputRHS takes the same tensor shapes as the matmul itself, and returns a tensor of
the same shape as given.

The MatMul option flag `inputRHSIsPreArranged` controls whether or not the pre-arrangement takes place, so it's easy to toggle.

Test Plan: `./test.sh poplibs`

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D7075