Re-arrange slice inputs to dynamic update functions to reduce vertex state (use supervisor rather than 2d vertices).
Summary:
In the forward pass with training, there are several separate variables which form the retained state, and as such we currently have to rearrange them into a single new tensor to dynamic update them into the retained sequence without the use of 2d vertices.
In the backward pass, the slice to dynamic update is input gradients which due to our combined input/output matmul are interleaved with prev step output gradients. This is unavoidable while we have this combined matmul.

We may be able to rearrange these more efficiently or avoid the rearrangement altogther depending on upcoming API changes but this is a quick fix for now.

Saves ~40MB of memory for our 8-layer LSTM benchmark with training, batch size 16, 1024 hidden units, 25 steps.

Test Plan: `./test.sh poplibs`

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D7111