Update LSTM API
Summary:
Add a outputFullSequence flag to LstmParams that controls whether to
return the output for the last timestep or just the last one. Use pointers
for optional results that are not always required - if they are null
we may be able to reduce the amount of state we need to store.

As part of this change I've split out the forward pass output and the
state retained for training into separate sequences. Although this results
in two dynamic updates in the forward pass it improves the efficiency of
the dynamic slice in the next layer in the forward pass.

We now store a shifted copy of the output sequence in the intermediates
tensor - this avoids the need to append the initial state to the start of
the the output sequence at the start of the backward pass at the cost
of memory in the forwards pass. This was done for simplicity - I
intend to investigate changing this back to using a shifted version
of the outputs to save memory.

The sequence of gradients produced in the backward pass now use a grouping
of 16 instead of being grouped based on the vector width. Again this was
done for simplicity - the best group size needs investigation

Reviewers: tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D7109