Optimize transposition of the output of the matrix multiply
Summary:
Use transpose vertices to rearrange the output so we have largish
blocks of elements that are laid out in the same order as the underlying
storage of the destination. The subsequent copy only has to copy large
blocks as a result, improving efficiency.

As part of this change I've move weightsPartialTranspose out of
Convolution.cpp and renamed it to partialTranspose. I've kept it as part
of popconv rather than moving it to popops since it is only optimized
for the group sizes you typically get in convolutions.

Fixes T4970

Reviewers: Eddy, tomm

Reviewed By: tomm

Maniphest Tasks: T4970

Differential Revision: https://phabricator.sourcevertex.net/D7065