Hoist accumulation across the batch for bias gradients out of the loop
Summary:
This change lets us base the layout of the accumulated bias gradients on
the layout of the intermediates, avoiding the need for exchange of these
intermediates in the loop. The cost of reducing across the batch is now
amortised over the sequence length. For a LSTM layer with
hidden=1024 --batch-size=16 the time taken up by bias reduction in each
iteration of the WU loop is reduced from 3539 cycles to 213 cycles.

Reviewers: tomm

Reviewed By: tomm

Subscribers: davel

Differential Revision: https://phabricator.sourcevertex.net/D7157