Add tile dithering to convolutions based on their parameters
Summary:
Fixes T4927.

This provides much better tile balance which lowers the peak tile memory and allows more networks to fit that are close to the memory limit.

Unfortunately the cost of this seems to be about an 11% increase in cycles when applying this to resnet 50, training, bs1:
|          | total memory | mean tile mem | max tile mem | compute cycles | exchange cycles | sync cycles
|----------|--------------|---------------|--------------|----------------|-----------------|-------------
| disabled | 254.5MB      | 208,447B      | 232,312B     | 1,617,887      | 416,936         | 890,433
| enabled  | 244.8MB      | 201,322B      | 235,574B     | 1,574,836      | 464,285         | 1,292,941

Test Plan: Added the new option to `single_conv_layer_random.py`

Reviewers: richard

Reviewed By: richard

Subscribers: davidn

Maniphest Tasks: T4927

Differential Revision: https://phabricator.sourcevertex.net/D7038